# 2026-cohort/quantum-solvers-equitable-mentorship-pairing-engine/src/openapi.yaml
openapi: 3.1.0
info:
  title: Braid API
  version: 0.1.0
  description: |
    Thin vertical slice: program lookup, application intake, magic-link auth,
    roster, matching runs with a fairness summary, and strands with messages.

    The published form schema is here too, so the apply screen renders what a
    coordinator built rather than a hardcoded set of questions.

    Deliberately excluded until the step that needs them: file upload, and the
    pair-level run actions (swap, lock, reject, re-run).
servers:
  - url: http://localhost:8000/v1
    description: Local FastAPI
  - url: https://api.braid.app/v1
    description: Production
security:
  - sessionCookie: []
...
paths:
...
  /orgs/{orgSlug}/programs/{programSlug}/waitlist:
    parameters:
      - $ref: "#/components/parameters/OrgSlug"
      - $ref: "#/components/parameters/ProgramSlug"
    post:
      operationId: joinWaitlist
      tags: [Public]
      summary: Ask to be told when the next round opens
      description: |
        The action a closed, full or not-yet-open programme offers instead of a
        dead end. Always returns 202, for the same reason the magic-link
        endpoint does: a different answer for a known address would make this an
        enumeration oracle.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                role:
                  $ref: "#/components/schemas/Role"
      responses:
        "202":
          description: Accepted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

        "429":
          $ref: "#/components/responses/TooManyRequests"
  /orgs/{orgSlug}/programs/{programSlug}/application-draft:
    parameters:
      - $ref: "#/components/parameters/OrgSlug"
      - $ref: "#/components/parameters/ProgramSlug"
    get:
      operationId: getApplicationDraft
      tags: [Public]
      summary: Read a saved draft
      security: []
      parameters:
        - name: draftId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationDraft"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: saveApplicationDraft
      tags: [Public]
      summary: Autosave a partly filled application
      description: |
        Upsert. Omit `draftId` on the first call and the server mints one; the
        client holds it and sends it back on every later save.

        The draft is reachable by whoever holds the id, which is the same
        exposure as `GET /applications/{applicationId}` and is acceptable for
        the same reason: it carries no credential and grants nothing beyond the
        answers already typed on that device.

        Not yet built: emailing a resume link, which architecture 3.2 lists as
        an action. Until it exists the id lives only in the browser that started
        the form, and the screen says so rather than implying otherwise.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationDraftSave"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationDraft"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Applications are closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /invites/{token}:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getInvite
      tags: [Public]
      summary: Read an invitation
      description: |
        Returns the invitation whatever state it is in, including expired. A
        404 for a spent token would send somebody to a dead end when the useful
        answer is that it lapsed and can be reissued.
      security: []
      responses:
        "200":
          description: The invitation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: respondToInvite
      tags: [Public]
      summary: Accept or decline an invitation
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteResponse"
      responses:
        "200":
          description: |
            Accepted. A session is opened, so somebody arriving from an email
            lands inside the programme rather than on a sign-in form.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteAccepted"
        "204":
          description: Declined
        "400":
          $ref: "#/components/responses/BadRequest"
        "410":
          description: The invitation has expired or was already used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"

        "429":
          $ref: "#/components/responses/TooManyRequests"
  /invites/{token}/reissue:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: requestNewInvite
      tags: [Public]
      summary: Ask the coordinator for a fresh invitation
      description: |
        What an expired invitation offers instead of a dead end. The coordinator
        is told; no new token is minted here.
      security: []
      responses:
        "202":
          description: The coordinator has been asked
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: braid_session
  parameters:
    OrgSlug:
      name: orgSlug
      in: path
      required: true
      schema:
        type: string
    ProgramSlug:
      name: programSlug
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: The request failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    NotFound:
      description: No such resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    TooManyRequests:
      description: Rate limited
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable identifier, e.g. `applications_closed`.
        message:
          type: string
          description: Sentence-case text safe to show a person.
        field:
          type: string
          description: Present when the failure belongs to one input.
    Role:
      type: string
      enum: [mentor, mentee]
    ParticipationStatus:
      type: string
      enum: [invited, applied, approved, waitlisted, removed]
...
    AnswerValue:
      description: One answer to one question.
      oneOf:
        - type: string
        - type: number
        - type: boolean
        - type: array
          items:
            type: string
    Provenance:
      type: string
      enum: [self, guided]
      description: |
        Where the words came from. `self` is typed unprompted; `guided` is the
        answer to an elicitation follow-up on the guided completion screen.

        Recorded because scoring free text rewards writing fluency, which tracks
        schooling and language background. Reports have to be able to ask
        whether coverage improved without the guided answers quietly becoming a
        ranking feature.
    AnswerInput:
      type: object
      required: [value, provenance]
      description: |
        One answer on the way in. There is no timestamp: the server stamps it,
        because a client-supplied time is not evidence of anything.
      properties:
        value:
          $ref: "#/components/schemas/AnswerValue"
        provenance:
          $ref: "#/components/schemas/Provenance"
    AnswerRecord:
      type: object
      required: [value, provenance, answeredAt]
      description: One stored answer, as read back.
      properties:
        value:
          $ref: "#/components/schemas/AnswerValue"
        provenance:
          $ref: "#/components/schemas/Provenance"
        answeredAt:
          type: string
          format: date-time
...
    ApplicationDraftSave:
      type: object
      required: [role, formVersionId, answers]
      properties:
        draftId:
          type: string
          format: uuid
          description: Omit on the first save. The server mints one and returns it.
        role:
          $ref: "#/components/schemas/Role"
        formVersionId:
          type: string
          format: uuid
        answers:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/AnswerInput"
    ApplicationDraft:
      type: object
      required: [draftId, role, formVersionId, answers, savedAt]
      properties:
        draftId:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/Role"
        formVersionId:
          type: string
          format: uuid
        answers:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/AnswerRecord"
        savedAt:
          type: string
          format: date-time
          description: |
            What the autosave indicator shows. Section 9: a timestamp, never a
            spinner — a spinner says "wait", and the applicant has nothing to
            wait for.
...
    Account:
      type: object
      required: [id, name, email, emailVerified]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        photoUrl:
          type: [string, "null"]
          format: uri
    ParticipationSummary:
      type: object
      required: [id, programId, programName, orgSlug, programSlug, role, status]
      properties:
        id:
          type: string
          format: uuid
        programId:
          type: string
          format: uuid
        programName:
          type: string
        organisationName:
          type: string
        orgSlug:
          type: string
        programSlug:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        status:
          $ref: "#/components/schemas/ParticipationStatus"
        isCoordinator:
          type: boolean
          default: false
    Session:
      type: object
      required: [account, participations]
      properties:
        account:
          $ref: "#/components/schemas/Account"
        participations:
          type: array
          items:
            $ref: "#/components/schemas/ParticipationSummary"
...
    InviteState:
      type: string
      enum: [pending, accepted, declined, expired]
    Invite:
      type: object
      required:
        [
          token,
          state,
          email,
          organisationName,
          programName,
          orgSlug,
          programSlug,
          role,
          invitedByName,
        ]
      properties:
        token:
          type: string
        state:
          $ref: "#/components/schemas/InviteState"
        email:
          type: string
          format: email
        organisationName:
          type: string
        programName:
          type: string
        orgSlug:
          type: string
        programSlug:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        invitedByName:
          type: string
        message:
          type: [string, "null"]
          description: An optional line from whoever sent it.
        expiresAt:
          type: [string, "null"]
          format: date-time
        hasAccount:
          type: boolean
          default: false
          description: |
            Whether the invited address already belongs to an account. Decides
            whether accepting joins an existing one or starts a new one, and the
            screen says which so nobody fears a duplicate.
    InviteResponse:
      type: object
      required: [accept]
      properties:
        accept:
          type: boolean
        name:
          type: [string, "null"]
          description: Required when accepting without an existing account.
    InviteAccepted:
      type: object
      required: [session, orgSlug, programSlug]
      properties:
        session:
          $ref: "#/components/schemas/Session"
        orgSlug:
          type: string
        programSlug:
          type: string
...